Time: 3 ms (18.08%) | Memory: 7.8 MB (69.00%) - LeetSync

+++ 278-first-bad-version/first-bad-version.cpp
@@ -4,9 +4,9 @@
 class Solution {
 public:
     int firstBadVersion(int n) {
-        long long left = 1, right = n;
+        long left = 1, right = n;
         while( left < right){
-            long long mid = (left + right) / 2;
+            long mid = (left + right) / 2;
             if(isBadVersion(mid)){
                 right = mid;
             } else {
